split up TestObject file into one for each class
