+++ modules/ObjectControl/CMakeLists.txt
@@ -43,6 +43,8 @@ add_executable(${OBJECT_CONTROL_TARGET}
 	${CMAKE_CURRENT_SOURCE_DIR}/src/main.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/src/testobject.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/src/objectlistener.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/src/objectconnection.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/src/channel.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/src/states/state.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/src/states/idle.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/src/states/initialized.cpp
