+++ modules/ObjectControl/src/testobject.cpp
@@ -11,21 +11,12 @@
 using namespace ATOS;
 using std::placeholders::_1;
 
-TestObject::TestObject(rclcpp::Logger log,
-		std::shared_ptr<ROSChannels::Path::Sub> pathSub,
-		std::shared_ptr<ROSChannels::Monitor::Pub> monrPub,
-		std::shared_ptr<ROSChannels::NavSatFix::Pub> navSatFixPub
-	) :
-	osiChannel(SOCK_STREAM, log),
-	comms(log),
-	//Loggable(log),
-	//pathSub(pathSub),
-	//monrPub(monrPub),
-	//navSatFixPub(navSatFixPub),
-	conf(log),
-	Node("name123")
+TestObject::TestObject(uint32_t id) :
+	Node("object_" + std::to_string(id)),
+	osiChannel(SOCK_STREAM, get_logger()),
+	comms(get_logger()),
+	conf(get_logger())
 	 {
-		auto id = 1;
 		pathSub = std::make_shared<ROSChannels::Path::Sub>(*this, id, std::bind(&TestObject::onPathMessage, this, _1, id));
 		monrPub = std::make_shared<ROSChannels::Monitor::Pub>(*this, id);
 		navSatFixPub = std::make_shared<ROSChannels::NavSatFix::Pub>(*this, id);
@@ -41,9 +32,7 @@ TestObject::TestObject(TestObject&& other) :
 	conf(other.conf),
 	lastMonitor(other.lastMonitor),
 	maxAllowedMonitorPeriod(other.maxAllowedMonitorPeriod),
-	//Loggable(other.get_logger())
 	rclcpp::Node(other.get_name())
-
 {
 	this->comms = other.comms;
 	this->osiChannel = other.osiChannel;
@@ -235,109 +224,6 @@ void TestObject::establishConnection(std::shared_future<void> stopRequest) {
 	}
 }
 
-void Channel::connect(
-		std::shared_future<void> stopRequest,
-		const std::chrono::milliseconds retryPeriod) {
-	char ipString[INET_ADDRSTRLEN];
-	std::stringstream errMsg;
-
-	if (inet_ntop(AF_INET, &this->addr.sin_addr, ipString, sizeof (ipString))
-			== nullptr) {
-		errMsg << "inet_ntop: " << strerror(errno);
-		throw std::invalid_argument(errMsg.str());
-	}
-
-	if (this->addr.sin_addr.s_addr == 0) {
-		errMsg << "Invalid connection IP specified: " << ipString;
-		throw std::invalid_argument(errMsg.str());
-	}
-
-	std::string type = "???";
-	if (this->channelType == SOCK_STREAM) {
-		type = "TCP";
-	}
-	else if (this->channelType == SOCK_DGRAM) {
-		type = "UDP";
-	}
-
-	this->socket = ::socket(AF_INET, this->channelType, 0);
-	if (this->socket < 0) {
-		errMsg << "Failed to open " << type << " socket: " << strerror(errno);
-		this->disconnect();
-		throw std::runtime_error(errMsg.str());
-	}
-
-	// Begin connection attempt
-	RCLCPP_INFO(get_logger(), "Attempting %s connection to %s:%u", type.c_str(), ipString,
-			   ntohs(this->addr.sin_port));
-
-	while (true) {
-		if (::connect(this->socket,
-					reinterpret_cast<struct sockaddr *>(&this->addr),
-					sizeof (this->addr)) == 0) {
-			break;
-		}
-		else {
-			RCLCPP_ERROR(get_logger(), "Failed %s connection attempt to %s:%u, retrying in %.3f s ...",
-					   type.c_str(), ipString, ntohs(this->addr.sin_port), retryPeriod.count() / 1000.0);
-			if (stopRequest.wait_for(retryPeriod)
-					!= std::future_status::timeout) {
-				errMsg << "Connection attempt interrupted";
-				throw std::runtime_error(errMsg.str());
-			}
-		}
-	}
-}
-
-void Channel::disconnect() {
-	if (this->socket != -1) {
-		if (shutdown(this->socket, SHUT_RDWR) == -1) {
-			RCLCPP_ERROR(get_logger(), "Socket shutdown: %s", strerror(errno));
-		}
-		if (close(this->socket) == -1) {
-			RCLCPP_ERROR(get_logger(), "Socket close: %s", strerror(errno));
-		}
-		this->socket = -1;
-	}
-}
-
-void ObjectConnection::connect(
-		std::shared_future<void> stopRequest,
-		const std::chrono::milliseconds retryPeriod) {
-	try {
-		this->cmd.connect(stopRequest, retryPeriod);
-		this->mntr.connect(stopRequest, retryPeriod);
-	} catch (std::runtime_error& e) {
-		this->disconnect();
-		throw std::runtime_error(std::string("Failed to establish ISO 22133 connection. Reason: \n") + e.what());
-	}
-
-	return;
-}
-
-
-bool ObjectConnection::isConnected() const {
-	if (!isValid()) {
-		return false;
-	}
-	pollfd fds[2];
-	fds[0].fd = mntr.socket;
-	fds[0].events = POLLIN | POLLOUT;
-	fds[1].fd = cmd.socket;
-	fds[1].events = POLLIN | POLLOUT;
-	return poll(fds, 2, 0) >= 0;
-}
-
-bool ObjectConnection::isValid() const {
-	return this->cmd.isValid() && this->mntr.isValid();
-}
-
-void ObjectConnection::disconnect() {
-	this->cmd.disconnect();
-	this->mntr.disconnect();
-	close(this->interruptionPipeFds[0]);
-	close(this->interruptionPipeFds[1]);
-}
 
 void TestObject::sendHeartbeat(
 		const ControlCenterStatusType ccStatus) {
@@ -400,11 +286,6 @@ void TestObject::sendRemoteControl(bool on) {
 	this->comms.cmd << (on ? OBJECT_COMMAND_REMOTE_CONTROL : OBJECT_COMMAND_DISARM);
 }
 
-//void TestObject::sendOsiGlobalObjectGroundTruth(
-//	GlobalObjectGroundTruth_t gogt) {
-	//this->comms.osi << llls;
-//}
-
 void TestObject::sendOsiData(
 		const OsiHandler::LocalObjectGroundTruth_t& osidata,
 		const std::string& projStr,
@@ -427,234 +308,4 @@ void TestObject::sendStart(std::chrono::system_clock::time_point startTime) {
 
 void TestObject::sendControlSignal(const ControlSignalPercentage::SharedPtr csp) {
 	this->comms.mntr << csp;
-}
-
-ISOMessageID Channel::pendingMessageType(bool awaitNext) {
-	auto result = recv(this->socket, this->receiveBuffer.data(), this->receiveBuffer.size(), (awaitNext ? 0 : MSG_DONTWAIT) | MSG_PEEK);
-	if (result < 0 && !awaitNext && (errno == EAGAIN || errno == EWOULDBLOCK)) {
-		return MESSAGE_ID_INVALID;
-	}
-	else if (result < 0) {
-		throw std::runtime_error(std::string("Failed to check pending message type (recv: ")
-									 + strerror(errno) + ")");
-	}
-	else if (result == 0) {
-		throw std::runtime_error("Connection reset by peer");
-	}
-	else {
-		ISOMessageID retval = getISOMessageType(this->receiveBuffer.data(), this->receiveBuffer.size(), false);
-		if (retval == MESSAGE_ID_INVALID) {
-			throw std::runtime_error("Non-ISO message received from " + this->remoteIP());
-		}
-		return retval;
-	}
-}
-
-ISOMessageID ObjectConnection::pendingMessageType(bool awaitNext) {
-	if (awaitNext) {
-		if (!isValid()) {
-			throw std::invalid_argument("Attempted to check pending message type for unconnected object");
-		}
-		fd_set fds;
-		FD_ZERO(&fds);
-		FD_SET(interruptionPipeFds[0], &fds);
-		FD_SET(mntr.socket, &fds);
-		FD_SET(cmd.socket, &fds);
-		auto result = select(std::max({mntr.socket,cmd.socket,interruptionPipeFds[0]})+1,
-							 &fds, nullptr, nullptr, nullptr);
-		if (result < 0) {
-			throw std::runtime_error(std::string("Failed socket operation (select: ") + strerror(errno) + ")"); // TODO clearer
-		}
-		else if (!isValid()) {
-			throw std::invalid_argument("Connection invalidated during select call");
-		}
-		else if (FD_ISSET(interruptionPipeFds[0], &fds)){
-			close(interruptionPipeFds[0]);
-			throw std::range_error("Select call was interrupted");
-		}
-		else if (FD_ISSET(mntr.socket, &fds)) {
-			return this->mntr.pendingMessageType();
-		}
-		else if (FD_ISSET(cmd.socket, &fds)) {
-			return this->cmd.pendingMessageType();
-		}
-		throw std::logic_error("Call to select returned unexpectedly: " + std::to_string(result));
-	}
-	else {
-		auto retval = this->mntr.pendingMessageType();
-		return retval != MESSAGE_ID_INVALID ? retval : this->cmd.pendingMessageType();
-	}
-}
-
-Channel& operator<<(Channel& chnl, const HeabMessageDataType& heartbeat) {
-	auto nBytes = encodeHEABMessage(&heartbeat.dataTimestamp, heartbeat.controlCenterStatus,
-									chnl.transmitBuffer.data(), chnl.transmitBuffer.size(), false);
-	if (nBytes < 0) {
-		throw std::invalid_argument(std::string("Failed to encode HEAB message: ") + strerror(errno));
-	}
-	nBytes = send(chnl.socket, chnl.transmitBuffer.data(), static_cast<size_t>(nBytes), 0);
-	if (nBytes < 0) {
-		throw std::runtime_error(std::string("Failed to send HEAB: ") + strerror(errno));
-	}
-	return chnl;
-}
-
-Channel& operator<<(Channel& chnl, const ObjectSettingsType& settings) {
-	auto nBytes = encodeOSEMMessage(&settings, chnl.transmitBuffer.data(), chnl.transmitBuffer.size(), false);
-	if (nBytes < 0) {
-		throw std::invalid_argument(std::string("Failed to encode OSEM message: ") + strerror(errno));
-	}
-	nBytes = send(chnl.socket, chnl.transmitBuffer.data(), static_cast<size_t>(nBytes), 0);
-	if (nBytes < 0) {
-		throw std::runtime_error(std::string("Failed to send OSEM: ") + strerror(errno));
-	}
-	return chnl;
-}
-
-Channel& operator<<(Channel& chnl, const Trajectory& traj) {
-	ssize_t nBytes;
-
-	// TRAJ header
-	nBytes = encodeTRAJMessageHeader(
-				traj.id, TRAJECTORY_INFO_RELATIVE_TO_ORIGIN, traj.name.c_str(),traj.name.length(),
-				static_cast<uint32_t>(traj.points.size()), chnl.transmitBuffer.data(),
-				chnl.transmitBuffer.size(), false);
-	if (nBytes < 0) {
-		throw std::invalid_argument(std::string("Failed to encode TRAJ message: ") + strerror(errno));
-	}
-	nBytes = send(chnl.socket, chnl.transmitBuffer.data(), static_cast<size_t>(nBytes), 0);
-	if (nBytes < 0) {
-		throw std::runtime_error(std::string("Failed to send TRAJ message header: ") + strerror(errno));
-	}
-
-	// TRAJ points
-	for (const auto& pt : traj.points) {
-		struct timeval relTime;
-		CartesianPosition pos = pt.getISOPosition();
-		SpeedType spd = pt.getISOVelocity();
-		AccelerationType acc = pt.getISOAcceleration();
-
-		relTime = to_timeval(pt.getTime());
-
-		nBytes = encodeTRAJMessagePoint(&relTime, pos, spd, acc, static_cast<float>(pt.getCurvature()),
-										chnl.transmitBuffer.data(), chnl.transmitBuffer.size(), false);
-		if (nBytes < 0) {
-			// TODO what to do here?
-			throw std::invalid_argument(std::string("Failed to encode TRAJ message point: ") + strerror(errno));
-		}
-		nBytes = send(chnl.socket, chnl.transmitBuffer.data(), static_cast<size_t>(nBytes), 0);
-
-		if (nBytes < 0) {
-			// TODO what to do here?
-			throw std::runtime_error(std::string("Failed to send TRAJ message point: ") + strerror(errno));
-		}
-	}
-
-	// TRAJ footer
-	nBytes = encodeTRAJMessageFooter(chnl.transmitBuffer.data(), chnl.transmitBuffer.size(), false);
-	if (nBytes < 0) {
-		throw std::invalid_argument(std::string("Failed to encode TRAJ message footer: ") + strerror(errno));
-	}
-	nBytes = send(chnl.socket, chnl.transmitBuffer.data(), static_cast<size_t>(nBytes), 0);
-	if (nBytes < 0) {
-		throw std::runtime_error(std::string("Failed to send TRAJ message footer: ") + strerror(errno));
-	}
-	return chnl;
-}
-
-Channel& operator<<(Channel& chnl, const ObjectCommandType& cmd) {
-	auto nBytes = encodeOSTMMessage(cmd, chnl.transmitBuffer.data(), chnl.transmitBuffer.size(), false);
-	if (nBytes < 0) {
-		throw std::invalid_argument(std::string("Failed to encode OSTM message: ") + strerror(errno));
-	}
-	nBytes = send(chnl.socket, chnl.transmitBuffer.data(), static_cast<size_t>(nBytes), 0);
-	if (nBytes < 0) {
-		throw std::runtime_error(std::string("Failed to send OSTM: ") + strerror(errno));
-	}
-	return chnl;
-}
-
-Channel& operator<<(Channel& chnl, const StartMessageType& strt) {
-	auto nBytes = encodeSTRTMessage(&strt, chnl.transmitBuffer.data(), chnl.transmitBuffer.size(), false);
-	if (nBytes < 0) {
-		throw std::invalid_argument(std::string("Failed to encode STRT message: ") + strerror(errno));
-	}
-
-	nBytes = send(chnl.socket, chnl.transmitBuffer.data(), static_cast<size_t>(nBytes), 0);
-	if (nBytes < 0) {
-		throw std::runtime_error(std::string("Failed to send STRT: ") + strerror(errno));
-	}
-	return chnl;
-}
-
-Channel& operator>>(Channel& chnl, MonitorMessage& monitor) {
-	if (chnl.pendingMessageType() == MESSAGE_ID_MONR) {
-		struct timeval tv;
-		TimeSetToCurrentSystemTime(&tv);
-		auto nBytes = decodeMONRMessage(chnl.receiveBuffer.data(), chnl.receiveBuffer.size(), tv,
-										&monitor.first, &monitor.second, false);
-		if (nBytes < 0) {
-			throw std::invalid_argument("Failed to decode MONR message");
-		}
-		else {
-			nBytes = recv(chnl.socket, chnl.receiveBuffer.data(), static_cast<size_t>(nBytes), 0);
-			if (nBytes <= 0) {
-				throw std::runtime_error("Unable to clear from socket buffer");
-			}
-		}
-	}
-	return chnl;
-}
-
-Channel& operator>>(Channel& chnl, ObjectPropertiesType& prop) {
-	if (chnl.pendingMessageType() == MESSAGE_ID_VENDOR_SPECIFIC_ASTAZERO_OPRO) {
-		auto nBytes = decodeOPROMessage(&prop, chnl.receiveBuffer.data(), chnl.receiveBuffer.size(), false);
-		if (nBytes < 0) {
-			throw std::invalid_argument(strerror(errno));
-		}
-		else {
-			nBytes = recv(chnl.socket, chnl.receiveBuffer.data(), static_cast<size_t>(nBytes), 0);
-			if (nBytes <= 0) {
-				throw std::runtime_error("Unable to clear from socket buffer");
-			}
-		}
-	}
-	return chnl;
-}
-
-Channel& operator<<(Channel& chnl, const ControlSignalPercentage::SharedPtr csp) {
-	RemoteControlManoeuvreMessageType rcmm;
-	rcmm.command = MANOEUVRE_NONE;
-	rcmm.isThrottleManoeuvreValid = true;
-	rcmm.isBrakeManoeuvreValid = true;
-	rcmm.isSteeringManoeuvreValid = true;
-	rcmm.throttleUnit = ISO_UNIT_TYPE_THROTTLE_PERCENTAGE;
-	rcmm.brakeUnit = ISO_UNIT_TYPE_BRAKE_PERCENTAGE;
-	rcmm.steeringUnit = ISO_UNIT_TYPE_STEERING_PERCENTAGE;
-	rcmm.throttleManoeuvre.pct = csp->throttle;
-	rcmm.brakeManoeuvre.pct = csp->brake;
-	rcmm.steeringManoeuvre.pct = csp->steering_angle;
-	auto nBytes = encodeRCMMMessage(&rcmm, chnl.transmitBuffer.data(), chnl.transmitBuffer.size(), false);
-	if (nBytes < 0) {
-		throw std::invalid_argument(std::string("Failed to encode RCM message: ") + strerror(errno));
-	}
-	nBytes = send(chnl.socket, chnl.transmitBuffer.data(), static_cast<size_t>(nBytes), 0);
-	if (nBytes < 0) {
-		throw std::runtime_error(std::string("Failed to send RCM: ") + strerror(errno));
-	}
-	return chnl;
-}
-
-Channel& operator<<(Channel& chnl, const std::vector<char>& data) {
-	auto nBytes = send(chnl.socket, data.data(), data.size(), 0);
-	if (nBytes < 0) {
-		throw std::runtime_error(std::string("Failed to send raw data: ") + strerror(errno));
-	}
-	return chnl;
-}
-
-std::string Channel::remoteIP() const {
-	char ipString[INET_ADDRSTRLEN];
-	inet_ntop(AF_INET, &this->addr.sin_addr, ipString, sizeof (ipString));
-	return std::string(ipString);
 }
